[FLINK-15955] Rename RemoteFunction* -> GrpcFunction*
